@@ -1,8 +1,8 @@
 name: launch_review_example
 description: Demonstrates how to use the launch_review plugin.
 
-environment:                                                      
-  sdk: '>=2.10.0 <3.0.0' 
+environment:
+  sdk: ">=2.17.1 <3.0.0"
 
 dependencies:
   flutter:
@@ -10,7 +10,7 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.0
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
   flutter_test:
