@@ -20,6 +20,6 @@ flutter:
       ios:
         pluginClass: LaunchReviewPlugin
 
-environment:  
-  sdk: ">=2.12.0 <3.0.0"
+environment:
+  sdk: ">=2.17.1 <3.0.0"
   flutter: ">=1.20.0"
